Updated QtRD-15810/qtcreator-super Project: Comm-15470/qtquickdesigner  264a1e27e7efb8cdfd381c1c9b9ec5e17a01601f
Add isoiconconverter tool.

This is a utility to convert the EPS files that we get from the
ISO 7000 package to SVG files via inkscape, and then DAT QPicture
files.

It should be built and used manually whenever an update of the
icons is deemed necessary, which shouldn't be that often (the ISO
organization periodically sends out updates to files).

Change-Id: I0faaab2fc31be312122e5ed6e4cd642a2ca46e9c
Reviewed-by: Thomas Hartmann <thomas.hartmann@theqtcompany.com>