Updated QtRD-15810/qtcreator-super Project: Comm-15470/qtquickdesigner  a18e2c9644420140897dfe3f15c698e7cf1d5963
Implement proper automatic layouting

If the user wants to layout items in a GridLayout
using the context menu we now provide a nice heuristic
which usually does what the user wants.

We calcualte the "minimal" grid, spanning and spacer items, so that
the GridLayout is very close to the absolute layout before.

Change-Id: Ic8d67482f97489a865ffe954c31622ca44b4edd0
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>