Updated QtRD-15810/qtcreator-super Project: Comm-15470/qtquickdesigner  c0a7c6336d4d3b3b014ee02443f9e21c6976113f
Add IsoIconBrowserPlugin

This plugin enables enterprise Creator users to easily browse through
the ~3000 ISO 7000 icons (which are added in a separate patch).

It provides sorting and filtering by certain metadata fields.

The icons can be displayed with the new Picture control.

The icons are embedded into the plugin. When the user selects an icon,
it is added to the ISO_ICONS qmake variable in the user's project,
which in turn updates the iso-icons.qrc file that we add to the
project.

Change-Id: If65b134fc0aecac78d7f040fc72d419c7c97c34f
Reviewed-by: Thomas Hartmann <thomas.hartmann@theqtcompany.com>