Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  868225d24e4b063205ff95bccbe65ffcc8e771e9
Take the middle between consecutive stacks as start/end of events

Previously the information in a sample was interpreted to span from the
time when it's collected to the time when the next sample is collected.
This change moves these start and end times to the middle between the
times when the samples are collected and adds support for thread start/exit
events to determine the very first and very last samples of a thread. The
effect is that a "sample collected" event is roughly in the middle of the
events that belong to it, instead of their very beginning.

Change-Id: I352eb39a3616392c2541cc2f77b9ec5ac138e9c2
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>