Updated QtRD-15810/qtcreator-super Project: Comm-15470/qtquickdesigner  3ce0414f082d9e3378e57094543e2fd815798fa3
Set QML[2]_IMPORT_PATH environment variable when running qml[2]puppet

The qml[2]puppet can't use any QML plugin that are defined in your .pro file
(using QML_IMPORT_PATH) unless we specify it.

To implement this we need the puppet version and the model as a
memeber in the PuppetCreator.

Change-Id: I420e683ea83365a56ab9c3a731de1e773ab726b8
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>