Updated QtRD-15810/qtcreator-super Project: QtRD-15810/autotest-qtcreator-plugin  2c1a0119bd56be2e18250e4c08c1fb201f541885
Fix unexpected parser state when parsing finishes

While parsing test code it might happen that the user closes
all related widgets and so the parser falls into the state Disabled.
This was unexpected and did print a warning.
Now this case is somewhat expected and won't print a warning anymore.

Change-Id: I0082bdc31f6ed351cf1dbd7b34da2f5672aee3dd
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@theqtcompany.com>
Reviewed-by: Christian Stenger <christian.stenger@theqtcompany.com>