Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  fc8dbf345534984aca3bd17f9461e721e4ac69fb
Use adaptive buffering to deal with unsorted samples

Perf doesn't always send samples in chronological order. As we cannot
insert into the middle of a model we have to buffer the samples before
we insert. However, we don't know the buffer length in advance. To
overcome this we store all samples in a QList and whenever we detect
a time shift, we increase the buffer length so that a time shift of the
same size won't be a problem in the future. When we do this we have to
copy and shorten the respective model, though.

Change-Id: I86bf879519647a192ab84d9e59a4f7ab4041af92
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>