Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  767ed27999766f5b84348d536ac102f973c75927
Calculate event saturation based on real frequency values.

The frequency setting is not very useful as we don't know if the kernel
actually manages to sample at the given frequency. Quite often it
won't. So instead, we just calculate the sampling frequency from the
samples and timestamps we receive. Also, we need to consider the number
of concurrently running threads for the saturation as the kernel will
only sample at a fixed rate, no matter how many threads are running.

Change-Id: Ibd06cbc7f980c8ed262cf43dfe9c12fe2bf09df1
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>