Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  afdcfc17633497f92e9153a568ebcaf2cb71383c
Show labels for recorded time and processing delay

The events are often delivered in bursts, with long gaps between
them. By showing the current processing delay and the time span
actually recorded we give the user a better indication of what will
actually be shown when the stop button is pressed.

Also, this will later allow us to make the recording more precise
by comparing the "local" and "remote" timestamps and starting and
stopping the recording at the projected time the user "meant" when
pressing the start or stop buttons.

Change-Id: I1632feff2dbd8bfc41e95e0778a943a50f4cf716
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>