Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  74b6a7b0b5b4a197d811592f53394738d7e6fd85
Use signed integers for all timestamps and durations

This adds one cast when converting from perf to local format but avoids
potentially many nasty surprises when interacting with
TimelineZoomControl. For example the default invalid range of
PerfTimelineModelManager, (uint64_max, 0), was casted to (-1, 0) and
thus resulted in a valid range for TimelineZoomControl.

Change-Id: Ib48ef14706be0c7f9bbd9c4f69b71e4dc26e7832
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>