Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  91cf40d6f2af5effedaa0d6e79c57fc8c1a1531e
Clean up interaction between trace loading and recording

Traces are loaded in the background now, with a loading indicator
to show an estimate of the progress. All the UI controls are
enabled and disabled in the right places, so that you cannot load
multiple traces on top of one another anymore and you cannot load
traces while recording anymore. You can, however, cancel the
loading now, and interact with other parts of QtCreator while it
takes place.

Change-Id: Ia717186feab3af9d5151aef16969023a7de00f1d
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>