Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  389964aa9aa78e28a1a1312336fcf720bf5b30f1
Sort models by tid before inserting them.

This results in a stable ordering of threads in the timeline. The order
will stay the same across multiple recording sessions while the process
is running.

As child threads are likely to have a higher tid than parent threads it
is also likely that parent threads show up above child threads, which is
a desirable effect.

Change-Id: I62596d18385c09e60637b21bf440e37dbbe5d8ce
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>