Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  a6efbcdf81c3392391a7b26250a3af22663b3dd5
Instead of terminating the parser just close its write channel.

If we terminate the parser it will drop any pending data, which leads
to truncated traces. This is especially noticeable if there is very
little data to begin with and all of it remains in the perfparser's
output buffer.

The perfparser will terminate itself once it has read all data. In the
local desktop case this is triggered by the closing of the input
channel. In the boot2qt case it's triggered by the adb port forwarding
going down. When reading a trace file it terminates itself once the
file is fully processed. Thus there is no need to explicitly terminate
it.

Change-Id: Iaf2ee91e9cf4062646c1c896b46a34fa9a37b2ae
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>