Updated QtRD-15810/qtcreator-super Project: QtRD-15810/autotest-qtcreator-plugin  00ca77a560021c332b3d6f76efb79d082e930da0
Do not allow triggering multiple test runs at once

This was possible as triggering further test runs could happen
while the current test run was still building the project.
Additionally it's now possible to stop the build process (if any)
by hitting the Stop button of the test results pane as this is
part of the test run.

Change-Id: I38940b5b8f4ba9e373785921c04cbceaeb2e5acf
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>