Updated QtRD-15810/qtcreator-super Project: Comm-15470/perfprofiler  19ed2cb184a5b2e08ce39f3b48f40814ae5abbef
Buffer data if the parser starts after perf is already running

As we need to start the process to figure out the port to connect to
we cannot start the parser before starting the process in the general
case. If the process then starts producing data before the parser
is ready we need to buffer that data.

Change-Id: I79e2e8914b16363cbcbe26b7f9e1d6a4a2d4910e
Reviewed-by: Kari Hautamäki <kari.hautamaki@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>